requirements.txt
Werkzeug==1.0.0
requests==2.23.0
Flask_SocketIO==4.2.1
Flask_Cors==3.0.8
Flask==1.1.1
psutil
GPUtil==1.4.0
flask_mongoengine